fixes template w/o product

+++ services/web/server/src/simcore_service_webserver/products.py
@@ -12,6 +12,7 @@ At every request to this service API, a middleware discovers which product is th
 import logging
 import os.path
 from pathlib import Path
+from typing import Optional
 
 import aiofiles
 from aiohttp import web
@@ -85,31 +86,39 @@ async def get_product_template_path(request: web.Request, filename: str) -> Path
             raise ValueError(f"Missing template {template_name} for product")
         return content
 
-    product: Product = get_current_product(request)
+    def _safe_get_current_product(request: web.Request) -> Optional[Product]:
+        try:
+            product: Product = get_current_product(request)
+            return product
+        except KeyError:
+            return None
 
-    if template_name := product.get_template_name_for(filename):  # type: ignore
-        template_dir = request.app[APP_PRODUCTS_TEMPLATES_DIR_KEY]
-        template_path = template_dir / template_name
-        if not template_path.exists():
-            # cache
-            content = await _get_content(template_name)
-            try:
-                async with aiofiles.open(template_path, "wt") as fh:
-                    await fh.write(content)
-            except Exception:
-                # fails to write
-                if template_path.exists():
-                    template_path.unlink()
-                raise
+    # ---
+    if product := _safe_get_current_product(request):
+        if template_name := product.get_template_name_for(filename):  # type: ignore
+            template_dir = request.app[APP_PRODUCTS_TEMPLATES_DIR_KEY]
+            template_path = template_dir / template_name
+            if not template_path.exists():
+                # cache
+                content = await _get_content(template_name)
+                try:
+                    async with aiofiles.open(template_path, "wt") as fh:
+                        await fh.write(content)
+                except Exception:
+                    # fails to write
+                    if template_path.exists():
+                        template_path.unlink()
+                    raise
 
-        return template_path
+            return template_path
 
-    # check static resources under templates/
-    if (
-        template_path := _themed(f"templates/{product.name}", filename)
-    ) and template_path.exists():
-        return template_path
+        # check static resources under templates/
+        if (
+            template_path := _themed(f"templates/{product.name}", filename)
+        ) and template_path.exists():
+            return template_path
 
+    # If no product or template for product defined, we fall back to common templates
     common_template = _themed("templates/common", filename)
     if not common_template.exists():
         raise ValueError(f"{filename} is not part of the templates/common")
